refactor: 인증 결과 및 역할 변경 안내 클래스 분리
- EmailNotificationService => 변경 안내 이메일 전송 관련 클래스
