+++ src/main/java/org/cotato/csquiz/domain/auth/service/EmailVerificationService.java
@@ -1,12 +1,6 @@
 package org.cotato.csquiz.domain.auth.service;
 
-import static org.cotato.csquiz.common.util.EmailUtil.createOldMemberConversionEmailBody;
-import static org.cotato.csquiz.common.util.EmailUtil.createSignupApprovedMessageBody;
-import static org.cotato.csquiz.common.util.EmailUtil.createSignupRejectionMessageBody;
 import static org.cotato.csquiz.common.util.EmailUtil.getVerificationMessageBody;
-import static org.cotato.csquiz.domain.auth.constant.EmailConstants.CONVERSION_TO_OM_SUBJECT;
-import static org.cotato.csquiz.domain.auth.constant.EmailConstants.SIGNUP_REJECT_SUBJECT;
-import static org.cotato.csquiz.domain.auth.constant.EmailConstants.SIGNUP_SUCCESS_SUBJECT;
 
 import java.util.concurrent.ThreadLocalRandom;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +9,6 @@ import org.cotato.csquiz.common.email.EmailService;
 import org.cotato.csquiz.common.error.exception.AppException;
 import org.cotato.csquiz.common.error.ErrorCode;
 import org.cotato.csquiz.domain.auth.cache.EmailRedisRepository;
-import org.cotato.csquiz.domain.auth.entity.Member;
 import org.cotato.csquiz.domain.auth.enums.EmailType;
 import org.cotato.csquiz.domain.auth.utils.EmailFormValidator;
 import org.cotato.csquiz.domain.auth.cache.VerificationCodeRedisRepository;
@@ -50,33 +43,6 @@ public class EmailVerificationService {
         emailService.sendEmail(recipient, verificationMessage, subject);
     }
 
-    public void sendSignUpApprovedToEmail(Member recipientMember) {
-        emailFormValidator.validateEmailForm(recipientMember.getEmail());
-
-        String successMessage = createSignupApprovedMessageBody(recipientMember);
-
-        emailService.sendEmail(recipientMember.getEmail(), successMessage, SIGNUP_SUCCESS_SUBJECT);
-        log.info("가입 승인 완료 이메일 전송 완료");
-    }
-
-    public void sendSignupRejectionToEmail(Member recipientMember) {
-        emailFormValidator.validateEmailForm(recipientMember.getEmail());
-
-        String rejectMessage = createSignupRejectionMessageBody(recipientMember);
-
-        emailService.sendEmail(recipientMember.getEmail(), rejectMessage, SIGNUP_REJECT_SUBJECT);
-        log.info("가입 승인 거절 이메일 전송 완료");
-    }
-
-    public void sendOldMemberConversionToEmail(Member recipientMember) {
-        emailFormValidator.validateEmailForm(recipientMember.getEmail());
-
-        String conversionMessageBody = createOldMemberConversionEmailBody(recipientMember);
-
-        emailService.sendEmail(recipientMember.getEmail(), conversionMessageBody, CONVERSION_TO_OM_SUBJECT);
-        log.info("OM 전환 이메일 전송 완료");
-    }
-
     private String getVerificationCode() {
         final ThreadLocalRandom random = ThreadLocalRandom.current();
         StringBuilder builder = new StringBuilder();
@@ -85,7 +51,6 @@ public class EmailVerificationService {
         }
         return String.valueOf(builder);
     }
-
 
     public void verifyCode(EmailType type, String email, String code) {
         String savedVerificationCode = verificationCodeRedisRepository.getByEmail(type, email);
